Refactor data loading and validation logic
Replaced the `create_dataloaders` function with a more robust version using the `TBDataLoader` class and updated tests to match the new implementation. Consolidated utility functions like `validate_eval_set` into appropriate modules and removed deprecated code to streamline functionality.
